:recycle: Replace phase-branch heuristics with an intercept test
The global 2*pi branch for the MCPC-3D-S phase offset was picked by a
cascade of five hardcoded field-magnitude thresholds, with a --wrap-limit
escape hatch for when the cascade made things worse. Replace it with a
criterion that has no tuned constants.

A branch error of M wraps shifts the phase offset by wrap(2*pi*k*M) with
k = TE0/dTE, and that lands as the same additive constant on every echo.
Fitting a line through the two echoes and reading its intercept measures
that constant directly: the correct branch extrapolates through the
origin, a wrong one does not. The rejection scale is analytic --
|wrap(2*pi*k)| -- so nothing is fit to data.

- Add _evaluate_branch / _branch_intercept_step / _select_branch; drop
  _select_branch_cascade, FMAP_PROPORTION_HEURISTIC,
  FMAP_AMBIGUIOUS_HEURISTIC, BRANCH_SELECTION_MODES and
  BRANCH_CONSISTENCY_RATIO.
- Act only when exactly one candidate is consistent. On a tie the
  branches are a genuine alias that no phase statistic can resolve, so
  defer to ROMEO's correct_global rather than guess: a measured
  centre-frequency sweep put a smallest-|field| tiebreaker at 12% better
  and 12% worse, including cases where correct_global had been right.
- Return 0 when TE0/dTE is an integer, where the branch cannot move the
  offset at all.
- Drop wrap_limit from medic(), unwrap_and_compute_field_maps() and the
  wk-medic / wk-unwrap-phase CLIs. BREAKING: --wrap-limit is gone; it
  existed to disable the heuristics this commit removes.
- Correct two annotations the new tests exposed: echo times are numpy
  scalars at every call site, not arrays, and rescale_phase's min/max
  are floats, not ints.
- Add 13 tests covering the analytic step, the decision rule, and
  end-to-end recovery of an injected wrap on the bundled data.

Validated on 14090 frames across 58 runs from eight OpenNeuro datasets
spanning six TE0/dTE ratios (0.51-0.78). No frame selected a non-zero
branch and the largest branch-0 intercept was 1.9e-07 of the cutoff, so
the output is unchanged on all data tested. Fitting and non-fitting
branches separate by six to nine orders of magnitude when the rule does
fire, verified by injecting known wraps.

+++ warpkit/utilities.py
@@ -141,7 +141,7 @@ def corr2_coeff(a: npt.NDArray, b: npt.NDArray) -> npt.NDArray:
 
 
 def rescale_phase(
-    data: npt.NDArray[Any], min: int = -4096, max: int = 4096
+    data: npt.NDArray[Any], min: float = -4096, max: float = 4096
 ) -> npt.NDArray[Any]:
     """Rescale phase data to [-pi, pi]
